refactor #118: move toggle button to atoms folder
