+++ components/scoreboardSelector/index.tsx
@@ -4,7 +4,7 @@ import { useTheme } from '@/hooks/useTheme'
 import { Clock } from '../icons/clock'
 import { Stats } from '../icons/stats'
 import { PageContainer } from '../pageContainer'
-import Toggle from '../toggle'
+import Toggle from '../atoms/toggle'
 import { useRouter } from 'next/navigation'
 
 export default function ScoreboardSelector({
